convVar_k200.R
this program reads the *.topic.frequency files for each dataset gets the top frequency for x=1..50, and merges them up together 
for them to be analyzed later with a regression model. date. Jan 2017 (This version considers only the first 200 topics and filters out then 0.75 of the data, this is for improving the fit)